+++ ajax/users.php
@@ -1,12 +1,18 @@
 <?
-$PWD = "../";
-require_once($PWD."include/common.php");
+$NPADMIN_PATH = "../";
+require_once($NPADMIN_PATH."include/common.php");
 
 header("Cache-Control: no-cache");
 header("Pragma: no-cache");
 header("Expires: Mon, 01 Jan 2000 01:00:00 GMT");
 
 $_POST = NP_UTF8_decode($_POST);
+
+$authClass = npadmin_setting("NP-ADMIN", "AUTH");
+$authenticator = null;
+if ($authClass != null) {
+	$authenticator = new $authClass;
+}
 
 foreach ($_POST as $k => $v) {
    if ($v === "null")
@@ -43,26 +49,37 @@ if (array_key_exists("op", $_POST) && ($_POST['op'] == "login" || $_POST['op'] =
    } else if ($_POST['op'] == "delete") {
       $list = split(",", $_POST['list']);
       foreach ($list as $id) {
-         $user = new User();
-         $user->user = $id;
-         if (!$user->delete()) {
-            echo "ERROR: Unable to delete user '".$id."'";
-            return;
+         if ($id != null) {
+            $user = new User();
+            $user->user = $id;
+            if (!$user->delete()) {
+               echo "ERROR: Unable to delete user '".$id."'";
+               return;
+            }
          }
       }
       echo "OK";
       
    } else if ($_POST['op'] == "listAssignedGroups") {
-      $groups = array();
-      
-      $groups = $ddbb->executeSelectQuery("SELECT g.".$ddbb->getMapping('Group','groupName')." AS group_name FROM ".$ddbb->getTable("User")." u, ".$ddbb->getTable("Group")." g, ".$ddbb->getTable("UserGroup")." ug WHERE u.".$ddbb->getMapping('User','user')." = ug.".$ddbb->getMapping('UserGroup','user')." AND ug.".$ddbb->getMapping('UserGroup','groupName')." = g.".$ddbb->getMapping('Group','groupName')." AND u.".$ddbb->getMapping('User','user')." = '".$_POST['user']."' ORDER BY 1");
+
+      $groups = $ddbb->executeSelectQuery("SELECT ug.group_name AS group_name FROM  ".$ddbb->getTable("UserGroup")." ug WHERE ug.user = '".$_POST['user']."' ORDER BY 1");
+      if ($authenticator != null) {
+	      if ($groups == null)
+		      $groups = array();
+	      $groups = array_merge($groups, $authenticator->listAssignedGroups($_POST['user']));
+      }
 
       echo NP_json_encode($groups);
 
    } else if ($_POST['op'] == "listUnassignedGroups") {
-      $groups = array();
-      
-      $groups = $ddbb->executeSelectQuery("SELECT ".$ddbb->getMapping('Group','groupName')." FROM ".$ddbb->getTable("Group")." WHERE ".$ddbb->getMapping('Group','groupName')." NOT IN (SELECT g.".$ddbb->getMapping('Group','groupName')." AS group_name FROM ".$ddbb->getTable("User")." u, ".$ddbb->getTable("Group")." g, ".$ddbb->getTable("UserGroup")." ug WHERE u.".$ddbb->getMapping('User','user')." = ug.".$ddbb->getMapping('UserGroup','user')." AND ug.".$ddbb->getMapping('UserGroup','groupName')." = g.".$ddbb->getMapping('Group','groupName')." AND u.".$ddbb->getMapping('User','user')." = '".$_POST['user']."') ORDER BY 1");
+
+      $groups = $ddbb->executeSelectQuery("SELECT group_name FROM ".$ddbb->getTable("Group")." WHERE group_name NOT IN (SELECT group_name FROM ".$ddbb->getTable("UserGroup")." WHERE user = '".$_POST['user']."') ORDER BY 1");
+
+      if ($authenticator != null) {
+	      if ($groups == null)
+		      $groups = array();
+	      $groups = array_merge($groups, $authenticator->listUnasssignedGroups($_POST['user']));
+      }
 
       echo NP_json_encode($groups);
       
@@ -72,11 +89,19 @@ if (array_key_exists("op", $_POST) && ($_POST['op'] == "login" || $_POST['op'] =
       
       $sql = "DELETE FROM ".$ddbb->getTable('UserGroup')." WHERE ".$ddbb->getMapping('UserGroup','user')." = ".NP_DDBB::encodeSQLValue($user, $ddbb->getType('UserGroup','user')); 
       $ddbb->executeDeleteQuery($sql);
-      
+
       foreach ($groups as $group_name) {
          if ($group_name != "") {
-            $ug = new UserGroup(array("group_name" => $group_name, "user" => $user));
-            $ug->store();
+      		$is_authenticator_group = false;
+      		if ($authenticator != null) {
+      		   if (strstr($group_name, $authenticator->prefix) !== false) {
+      			   $is_authenticator_group = true;
+      		   }
+      		}
+      		if (!$is_authenticator_group) {
+      		   $ug = new UserGroup(array("group_name" => $group_name, "user" => $user));
+      		   $ug->store();
+      		}
          }
       }
       
@@ -87,14 +112,11 @@ if (array_key_exists("op", $_POST) && ($_POST['op'] == "login" || $_POST['op'] =
    }
 
    if ($returnList) {
-      $users = array();
+      $users = $ddbb->executeSelectQuery("SELECT * FROM ".$ddbb->getTable("User")." ORDER BY 1");
 
-      function createUserList($user) {
-         global $users;
-         $users[] = $user;
+      if ($authenticator != null) {
+         $users = array_merge($users, $authenticator->listUsers());
       }
-
-      $ddbb->executeSelectQuery("SELECT * FROM ".$ddbb->getTable("User")." ORDER BY 1", "createUserList");
 
       echo NP_json_encode($users); 
    }
